nvme

+++ script/qb-nox-remove.sh
@@ -8,7 +8,19 @@
 username=$1
 
 #判断文件是否存在
-if [[ ! -f "/usr/bin/qb-nox-static-438-lt1214" ]]; then
+if [[ -f "/usr/bin/qb-nox-static-438-lt1214" ]]; then
+  # 4.3.8
+  echo -e "\033[36m ================= 删除qb-nox 4.3.8 ================= \033[0m"
+  systemctl stop qb-nox-static-438-lt1214@$username
+  rm -rf /etc/systemd/system/qb-nox-static-438-lt1214@.service
+  rm -rf /usr/bin/qb-nox-static-438-lt1214
+  systemctl disable qb-nox-static-438-lt1214@$username
+  rm -rf /home/$username/.config/qBittorrent/
+  rm -rf /home/$username/.local/share/
+  rm -rf /home/$username/.cache/qBittorrent/
+fi
+
+if [[ -f "/usr/bin/qb-nox-static-419-lt1114" ]]; then
   # 4.1.9
   echo -e "\033[36m ================= 删除qb-nox 4.1.9 ================= \033[0m"
   systemctl stop qb-nox-static-419-lt1114@$username
@@ -18,14 +30,16 @@ if [[ ! -f "/usr/bin/qb-nox-static-438-lt1214" ]]; then
   rm -rf /home/$username/.config/qBittorrent/
   rm -rf /home/$username/.local/
   rm -rf /home/$username/.cache/qBittorrent/
-else
-  # 4.3.8
-  echo -e "\033[36m ================= 删除qb-nox 4.3.8 ================= \033[0m"
-  systemctl stop qb-nox-static-438-lt1214@$username
-  rm -rf /etc/systemd/system/qb-nox-static-438-lt1214@.service
-  rm -rf /usr/bin/qb-nox-static-438-lt1214
-  systemctl disable qb-nox-static-438-lt1214@$username
+fi
+
+if [[ -f "/usr/bin/qb-nox-static-419-lt1114-nvme" ]]; then
+  # 4.1.9 nvme
+  echo -e "\033[36m ================= 删除qb-nox 4.1.9 ================= \033[0m"
+  systemctl stop qb-nox-static-419-lt1114-nvme@$username
+  rm -rf /etc/systemd/system/qb-nox-static-419-lt1114-nvme@.service
+  rm -rf /usr/bin/qb-nox-static-419-lt1114-nvme
+  systemctl disable qb-nox-static-419-lt1114-nvme@$username
   rm -rf /home/$username/.config/qBittorrent/
-  rm -rf /home/$username/.local/share/
+  rm -rf /home/$username/.local/
   rm -rf /home/$username/.cache/qBittorrent/
 fi
